test: add a common helper for test logger
