+++ internal/bpf/cgtracker_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/cilium/ebpf"
 	"github.com/rancher-sandbox/runtime-enforcer/internal/cgroups"
+	"github.com/rancher-sandbox/runtime-enforcer/internal/testutil"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sys/unix"
 )
@@ -26,7 +27,7 @@ func TestUpdateCgTrackerMap(t *testing.T) {
 	expectedMap := map[uint64]uint64{
 		cgroup1: cgroup1,
 	}
-	err = updateCgTrackerMap(newTestLogger(t), cgTrackerMap, cgroup1, "")
+	err = updateCgTrackerMap(testutil.NewTestLogger(t), cgTrackerMap, cgroup1, "")
 	require.NoError(t, err)
 	require.Equal(t, expectedMap, dumpMap(cgTrackerMap))
 
@@ -64,7 +65,7 @@ func TestUpdateCgTrackerMap(t *testing.T) {
 		expectedNestedCgroup3: cgroup1,
 	}
 
-	err = updateCgTrackerMap(newTestLogger(t), cgTrackerMap, cgroup1, tempDir)
+	err = updateCgTrackerMap(testutil.NewTestLogger(t), cgTrackerMap, cgroup1, tempDir)
 	require.NoError(t, err)
 	require.Equal(t, expectedMap, dumpMap(cgTrackerMap))
 }
